requirements_dev.txt
invoke==2.2.0
twine==5.0.0
wheel==0.42.0
zest.releaser==9.1.3

# Bare minimum requirements to provide us similar env to Cookiecutter Django.
dj-rest-auth==5.0.2
django-allauth==0.61.1
django-cors-headers==4.3.1
django==4.2.9
djangorestframework-simplejwt==5.3.1
djangorestframework==3.14.0
drf-spectacular==0.27.1
django-perm-filter==0.2.2
django-auditlog==2.3.0
django-celery-results==2.5.1
django-filter==23.5

##
pip-freezer==0.3.5
